Update README.md
- /combine [hand] [potion] Command
- /gravity [off/low/medium/high] Command
- /blink [distance] Command
- /itemroll [auto/once] Command
- /infinite [water/lava] Command
- /boost [enable/disable] Command
- /border [remove/square] [size] Command
- /rtp Command (Random tp within 5000 radius)

ItemRoll is a simple early stage. Can be modified.
Infinite water/lava is a bit buggy when spamming...
Simply place it not as fast to get it to work...
Combine may encounter issues; will fix along the way.
Created a config file; basic for now but will be better.
Added AutoLongDay.java into admin folder.